Remove Module but insure only rescuing safe errors Needs test refactor and fixing.
