+++ lib/stoplight/light/runnable.rb
@@ -52,15 +52,27 @@ module Stoplight
         failures = clear_failures
         on_success.call(failures) if on_success
         result
-      rescue error_handler => error
+      rescue Default::AllExceptionsExceptOnesWeMustNotRescue => error
         handle_error(error, on_failure)
       end
 
+      class ErrorHandler # rubocop:disable Style/Documentation
+        attr_reader :handle_error
+
+        def handle(error)
+          @handle_error = error
+        end
+      end
+
       def handle_error(error, on_failure)
-        size = record_failure(error)
-        on_failure.call(size, error) if on_failure
-        raise error unless fallback
-        fallback.call(error)
+        handler = ErrorHandler.new
+        error_handler.call(error, handler)
+        if handler.handle_error == error
+          size = record_failure(error)
+          on_failure.call(size, error) if on_failure
+          raise error unless fallback
+          fallback.call(error)
+        end
       end
 
       def clear_failures
